Remove a dead function, and make every dependency explicit
acceptAliasDigest had no caller and could not have the one its comment imagined:
wallets sign typed data, not bare digests. The encoding it was meant to pin is
already proven end to end — e2e-live builds a signature with the SDK's own
acceptAlias and submits it, which fails if the two disagree. That is a stronger
check than digest equality, and it does not cost permanent bytecode in a contract
that cannot be changed.

Swept the rest of the public surface for the same thing. Everything else with no
callers is inherited ERC-721 and EIP-712 (getApproved, tokenURI, supportsInterface,
eip712Domain) which cannot be removed, or getters that exist so a reader can
verify a wiring — transactVerifier, registryRootSeenAt, the level constants. Those
stay.

Then a stricter solhint: 90 warnings down to 46, none of them errors. Thirty
global imports became named ones, and the first compile after that failed — which
was the point. HaliasRegistry had been using PoseidonT4 and FIELD_PRIME that
arrived transitively through SMTRegistry's own global import, so nothing in the
file said where they came from. They are direct now.

The other fourteen were `uint[2]` mirroring the shape snarkjs generates. `uint` is
an alias for `uint256`, so making them explicit is type-identical and the verifier
call is unaffected.

Left as warnings deliberately: gas-increment-by-one and gas-strict-inequalities
suggest rewriting comparisons for savings the optimiser mostly already takes,
immutable-vars-naming would rename controller/pool/registry for style, and
gas-struct-packing cannot apply to TransactParams — its layout is the ABI a proof
commits to.

174 hardhat, 65 sdk, 108 e2e-live.

+++ packages/protocol/contracts/HaliasPool.sol
@@ -1,15 +1,15 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.28;
 
-import "./MerkleTreeWithHistory.sol";
-import "./base/Constants.sol";
-import "./interfaces/IHaliasPool.sol";
-import "./interfaces/IHaliasRegistry.sol";
-import "./interfaces/ITransactVerifier.sol";
+import { MerkleTreeWithHistory } from "./MerkleTreeWithHistory.sol";
+import { FIELD_PRIME, MAX_ABS_AMOUNT } from "./base/Constants.sol";
+import { TransactParams } from "./interfaces/IHaliasPool.sol";
+import { IHaliasRegistry } from "./interfaces/IHaliasRegistry.sol";
+import { ITransactVerifier } from "./interfaces/ITransactVerifier.sol";
 // SafeERC20 brings IERC20 with it; importing both only duplicates the symbol.
-import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
-import "@openzeppelin/contracts/utils/ReentrancyGuard.sol";
-import "@openzeppelin/contracts/utils/Address.sol";
+import { SafeERC20, IERC20 } from "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
+import { ReentrancyGuard } from "@openzeppelin/contracts/utils/ReentrancyGuard.sol";
+import { Address } from "@openzeppelin/contracts/utils/Address.sol";
 
 // ── Custom errors ─────────────────────────────────────────────────────────────
 
@@ -374,8 +374,8 @@ contract HaliasPool is MerkleTreeWithHistory, ReentrancyGuard {
         pubSignals[12] = uint256(p.outputCommitments[0]);
         pubSignals[13] = uint256(p.outputCommitments[1]);
 
-        (uint[2] memory pA, uint[2][2] memory pB, uint[2] memory pC) =
-            abi.decode(proof, (uint[2], uint[2][2], uint[2]));
+        (uint256[2] memory pA, uint256[2][2] memory pB, uint256[2] memory pC) =
+            abi.decode(proof, (uint256[2], uint256[2][2], uint256[2]));
 
         if (!transactVerifier.verifyProof(pA, pB, pC, pubSignals)) revert InvalidProof();
     }
